+++ src/changelog.md
@@ -13,6 +13,7 @@ Version 4.4 [*??.??.17*]
 * Startup user items are handled prior to Windows 7 in the same way as from Windows 8
   * Removed exporting as registry file but instead as .Startup and .CommonStartup
   * Removed "Delete backups after enabling" menu item
+* Removed "Default column width" menu item
 * Added "Quick-Search" to Autostart tab
 * Added "Execute" to popup menu
 * Improved "Open in Explorer"